packt_fwd_api/requirements.txt
blinker==1.9.0
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
tihan_sdv==0.1.5
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
